New render hot loop
Batch by mesh,shader only (materials need to be per mesh or part of instance buffers)
Simplified sort keys + submission loop
Added some stats

+++ playground/playground/graphics.h
@@ -56,5 +56,5 @@ private:
 	SDE::RenderSystem* m_renderSystem = nullptr;
 	SDE::JobSystem* m_jobSystem = nullptr;
 	Input::InputSystem* m_inputSystem = nullptr;
-	glm::vec2 m_windowSize;
+	glm::ivec2 m_windowSize;
 };
